+++ action_updater/client/listing.py
@@ -8,6 +8,10 @@ from action_updater.main import get_client
 
 def list_updaters(args, parser, extra, subparser):
     cli = get_client(quiet=args.quiet)
+
+    # Update config settings on the fly
+    cli.settings.update_params(args.config_params)
+
     items = [
         {"title": x.title, "identifier": name, "description": x.description}
         for name, x in cli.updaters.items()